Rewrite print_network_interfaces into NetworkInterfacesPrinter
